Update UART configuration in rtconfig.h and .config
This commit updates the UART configuration in the `rtconfig.h` and `.config` files. The `RT_CONSOLE_DEVICE_NAME` is changed from "uart1" to "uart3" in `rtconfig.h`, and the `CONFIG_RT_CONSOLE_DEVICE_NAME` is also changed from "uart1" to "uart3" in `.config`. Additionally, the UART configuration is updated in the `BSP_USING_UART` and `BSP_USING_UART3` sections of `.config`.